+++ crates/gdck-format/tests/style-guide/avoid_unnecessary_parentheses__good.gd
@@ -1,0 +1,3 @@
+func _ready():
+	if is_colliding():
+		queue_free()