+++ crates/gdck-format/tests/style-guide/functions_and_variables__neutral.expected.gd
@@ -1,0 +1,5 @@
+var particle_effect
+
+
+func load_level():
+	pass