+++ crates/gdck-format/tests/style-guide/functions_and_variables__neutral.gd
@@ -1,0 +1,3 @@
+var particle_effect
+func load_level():
+	pass