+++ crates/gdck-format/tests/style-guide/functions_and_variables__neutral_2.expected.gd
@@ -1,0 +1,5 @@
+var _counter = 0
+
+
+func _recalculate_path():
+	pass