+++ crates/gdck-format/tests/style-guide/indentation__good_3.gd
@@ -1,0 +1,18 @@
+var party = [
+	"Godot",
+	"Godette",
+	"Steve",
+]
+
+var character_dict = {
+	"Name": "Bob",
+	"Age": 27,
+	"Job": "Mechanic",
+}
+
+enum Tile {
+	BRICK,
+	FLOOR,
+	SPIKE,
+	TELEPORT,
+}