+++ crates/gdck-format/tests/style-guide/inferred_types__bad_2.gd
@@ -1,0 +1,3 @@
+# The compiler can't infer the exact type and will use Node
+# instead of ProgressBar.
+@onready var health_bar := get_node("UI/LifeBar")