+++ crates/gdck-format/tests/style-guide/inferred_types__good.gd
@@ -1,0 +1,5 @@
+# The type can be int or float, and thus should be stated explicitly.
+var health: int = 0
+
+# The type is clearly inferred as Vector3.
+var direction := Vector3(1, 2, 3)