+++ crates/gdck-format/tests/style-guide/methods_and_static_functions__neutral.gd
@@ -1,0 +1,29 @@
+func _init():
+	add_to_group("state_machine")
+
+
+func _ready():
+	state_changed.connect(_on_state_changed)
+	_state.enter()
+
+
+func _unhandled_input(event):
+	_state.unhandled_input(event)
+
+
+func transition_to(target_state_path, msg={}):
+	if not has_node(target_state_path):
+		return
+
+	var target_state = get_node(target_state_path)
+	assert(target_state.is_composite == false)
+
+	_state.exit()
+	self._state = target_state
+	_state.enter(msg)
+	Events.player_state_changed.emit(_state.name)
+
+
+func _on_state_changed(previous, new):
+	print("state changed")
+	state_changed.emit()