+++ crates/gdck-format/tests/style-guide/numbers__good.gd
@@ -1,0 +1,2 @@
+var float_number = 0.234
+var other_float_number = 13.0