+++ crates/gdck-format/tests/style-guide/whitespace__good.gd
@@ -1,0 +1,7 @@
+func _ready():
+	position.x = 5
+	position.y = target_position.y + 10
+	dict["key"] = 5
+	my_array = [4, 5, 6]
+	my_dictionary = { key = "value" }
+	print("foo")