+++ crates/gdck-format/tests/style-guide/whitespace__neutral.expected.gd
@@ -1,0 +1,4 @@
+func _ready():
+	x = 100
+	y = 100
+	velocity = 500